# internal/swagger/spec.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "durpdeploy REST API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/environments/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "environments"
        ],
        "summary": "Get an environment by ID.",
        "operationId": "getEnvironment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Environment represents a deployment target.",
            "schema": {
              "$ref": "#/definitions/Environment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "environments"
        ],
        "summary": "Update an environment.",
        "operationId": "updateEnvironment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EnvironmentRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Environment represents a deployment target.",
            "schema": {
              "$ref": "#/definitions/Environment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "409": {
            "description": "Conflict error response.",
            "schema": {
              "$ref": "#/definitions/ConflictError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "environments"
        ],
        "summary": "Delete an environment.",
        "operationId": "deleteEnvironment",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "EmptyResponse is returned for no-content endpoints.",
            "schema": {
              "$ref": "#/definitions/EmptyResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects/{id}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "projects"
        ],
        "summary": "Get a project by ID.",
        "operationId": "getProject",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ProjectResponse is the JSON shape for a project.",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "projects"
        ],
        "summary": "Update a project.",
        "operationId": "updateProject",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProjectRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ProjectResponse is the JSON shape for a project.",
            "schema": {
              "$ref": "#/definitions/ProjectResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "409": {
            "description": "Conflict error response.",
            "schema": {
              "$ref": "#/definitions/ConflictError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "projects"
        ],
        "summary": "Delete a project.",
        "operationId": "deleteProject",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "EmptyResponse is returned for no-content endpoints.",
            "schema": {
              "$ref": "#/definitions/EmptyResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects/{id}/releases/{relId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "releases"
        ],
        "summary": "Get a release with its variable snapshots.",
        "operationId": "getRelease",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "RelID",
            "description": "The release ID.",
            "name": "relId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ReleaseWithVariablesResponse is the detailed release payload.",
            "schema": {
              "$ref": "#/definitions/ReleaseWithVariablesResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects/{id}/schedules": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "schedules"
        ],
        "summary": "List scheduled deployments for a project.",
        "operationId": "listSchedules",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Limit",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "Offset",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "ScheduledDeploymentListResponse is a list of scheduled deployments.",
            "schema": {
              "$ref": "#/definitions/ScheduledDeploymentListResponse"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "schedules"
        ],
        "summary": "Create a scheduled deployment.",
        "operationId": "createSchedule",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ScheduleRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ScheduledDeployment is a cron-driven deployment configuration.",
            "schema": {
              "$ref": "#/definitions/ScheduledDeployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "422": {
            "description": "Validation error response.",
            "schema": {
              "$ref": "#/definitions/ValidationError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/projects/{id}/schedules/{schedId}/toggle": {
      "post": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "schedules"
        ],
        "summary": "Toggle active state of a scheduled deployment.",
        "operationId": "toggleSchedule",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "ID",
            "description": "The resource ID.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "SchedID",
            "description": "The schedule ID.",
            "name": "schedId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ScheduledDeployment is a cron-driven deployment configuration.",
            "schema": {
              "$ref": "#/definitions/ScheduledDeployment"
            }
          },
          "400": {
            "description": "Bad request error response.",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "401": {
            "description": "Unauthorized error response.",
            "schema": {
              "$ref": "#/definitions/UnauthorizedError"
            }
          },
          "404": {
            "description": "Not found error response.",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "500": {
            "description": "Server error response.",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "BadRequestError": {
      "type": "object",
      "title": "Bad request error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerBadRequestError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "ConflictError": {
      "type": "object",
      "title": "Conflict error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerConflictError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "EmptyResponse": {
      "type": "object",
      "title": "EmptyResponse is returned for no-content endpoints.",
      "x-go-name": "swaggerEmptyResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "Environment": {
      "type": "object",
      "title": "Environment represents a deployment target.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "tags": {
          "type": "string",
          "x-go-name": "Tags"
        }
      },
      "x-go-name": "swaggerEnvironment",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "EnvironmentRequest": {
      "type": "object",
      "title": "EnvironmentRequest is the body for create/update environment.",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "tags": {
          "type": "string",
          "x-go-name": "Tags"
        }
      },
      "x-go-name": "swaggerEnvironmentRequest",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "NotFoundError": {
      "type": "object",
      "title": "Not found error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerNotFoundError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ProjectResponse": {
      "type": "object",
      "title": "ProjectResponse is the JSON shape for a project.",
      "properties": {
        "created_at": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "CreatedAt"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "discord_webhook_url": {
          "type": "string",
          "x-go-name": "DiscordWebhookURL"
        },
        "gotify_token": {
          "type": "string",
          "x-go-name": "GotifyToken"
        },
        "gotify_url": {
          "type": "string",
          "x-go-name": "GotifyURL"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "lifecycle_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "LifecycleID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "notify_emails": {
          "type": "string",
          "x-go-name": "NotifyEmails"
        },
        "slack_webhook_url": {
          "type": "string",
          "x-go-name": "SlackWebhookURL"
        }
      },
      "x-go-name": "swaggerProjectResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "ScheduledDeploymentListResponse": {
      "type": "array",
      "title": "ScheduledDeploymentListResponse is a list of scheduled deployments.",
      "items": {
        "$ref": "#/definitions/ScheduledDeployment"
      },
      "x-go-name": "swaggerScheduledDeploymentListResponse",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
    "ServerError": {
      "type": "object",
      "title": "Server error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerServerError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
    "UnauthorizedError": {
      "type": "object",
      "title": "Unauthorized error response.",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-name": "swaggerUnauthorizedError",
      "x-go-package": "durpdeploy/internal/handler/api"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}